rename gaussian to gauss for consistency
