+++ src/rational/mat_q/sample/gauss.rs
